hw6/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cloud-sql-python-connector==1.4.2
cryptography==41.0.5
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
gcloud==0.18.3
google-api-core==2.12.0
google-auth==2.23.3
google-cloud-appengine-logging==1.3.2
google-cloud-audit-log==0.2.5
google-cloud-core==2.3.3
google-cloud-logging==3.8.0
google-cloud-pubsub==2.18.4
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpc-google-iam-v1==0.12.6
grpcio==1.59.0
grpcio-status==1.59.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numpy==1.26.1
oauth2client==4.1.3
pandas==2.1.2
proto-plus==1.22.3
protobuf==4.25.0rc2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.22
sympy==1.12
torch==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
waitress==2.1.2
Werkzeug==3.0.1
yarl==1.9.2
zope.interface==6.1
